+++ test/async.rb
@@ -8,10 +8,12 @@ end
 
 $docraptor = DocRaptor::DocApi.new
 
+output_file = "ruby-async.pdf"
+
 create_response = $docraptor.create_async_doc(
   test:             true,
   document_content: "<html><body>Hello from Ruby</body></html>",
-  name:             "ruby-async.pdf",
+  name:             output_file,
   document_type:    "pdf",
 )
 
@@ -22,4 +24,11 @@ status_response = nil
   sleep 1
 end
 
-$docraptor.get_async_doc(status_response.download_id)
+output_payload = $docraptor.get_async_doc(status_response.download_id)
+
+File.write(output_file, output_payload)
+output_type = `file -b #{output_file}`
+File.delete output_file
+
+raise "Output was not a PDF" unless output_type.start_with?("PDF")
+