+++ test/sync.rb
@@ -8,9 +8,17 @@ end
 
 $docraptor = DocRaptor::DocApi.new
 
-$docraptor.create_doc(
+output_file = "ruby-sync.pdf"
+
+output_payload = $docraptor.create_doc(
   test:             true,
   document_content: "<html><body>Hello from Ruby</body></html>",
-  name:             "ruby-sync.pdf",
+  name:             output_file,
   document_type:    "pdf",
 )
+
+File.write(output_file, output_payload)
+output_type = `file -b #{output_file}`
+File.delete output_file
+
+raise "Output was not a PDF" unless output_type.start_with?("PDF")